注文番号を表示するように修正
