+++ app/config/eccube/packages/eccube.yaml
@@ -99,3 +99,4 @@ parameters:
     eccube_forgot_mail_template_id: 6 #パスワードリセット
     eccube_reset_complete_mail_template_id: 7 #パスワードリマインダー
     eccube_shipping_notify_mail_template_id: 8 #出荷通知メール
+    eccube_order_code_format: '${dd}${yyyy}${number,5}${mm}'
