+++ src/Eccube/Resource/doctrine/import_csv/dtb_order.csv
@@ -1,1 +1,1 @@
-id,customer_id,country_id,pref_id,sex_id,job_id,payment_id,device_type_id,pre_order_id,code,message,name01,name02,kana01,kana02,company_name,email,tel01,tel02,tel03,fax01,fax02,fax03,zip01,zip02,zipcode,addr01,addr02,birth,subtotal,discount,delivery_fee_total,charge,tax,total,payment_total,payment_method,note,create_date,update_date,order_date,shipping_date,payment_date,order_status_id,discriminator_type
+id,customer_id,country_id,pref_id,sex_id,job_id,payment_id,device_type_id,pre_order_id,order_code,message,name01,name02,kana01,kana02,company_name,email,tel01,tel02,tel03,fax01,fax02,fax03,zip01,zip02,zipcode,addr01,addr02,birth,subtotal,discount,delivery_fee_total,charge,tax,total,payment_total,payment_method,note,create_date,update_date,order_date,shipping_date,payment_date,order_status_id,discriminator_type
