+++ src/Eccube/Resource/template/admin/Mail/point_notify.twig
@@ -12,7 +12,7 @@ file that was distributed with this source code.
 {{ Order.name01 }} {{ Order.name02 }} 様の保有ポイントがマイナスになっています。
 
 注文番号：{{ Order.id }}
-注文コード: {{ Order.code }}
+注文コード: {{ Order.order_code }}
 
 変更されるポイント：{{ changePoint|number_format }}
