+++ src/Eccube/Resource/template/admin/Order/edit.twig
@@ -135,7 +135,7 @@ file that was distributed with this source code.
                                     <div class="col-6">
                                         <div class="row mb-3">
                                             <div class="col-3"><span>{{ 'admin.order.edit.244'|trans }}</span></div>
-                                            <div class="col"><span class="mr-3">{{ Order.id }}</span></div>
+                                            <div class="col"><span class="mr-3">{{ Order.order_code }}</span></div>
                                         </div>
                                         <div class="row mb-3">
                                             <div class="col-3"><span>{{ 'admin.order.edit.246'|trans }}</span></div>
