+++ src/Eccube/Resource/template/admin/Shipping/order_item_prototype.twig
@@ -23,7 +23,7 @@ file that was distributed with this source code.
         </span>
     </td>
     <td class="align-middle">
-        <a href="{{ url('admin_order_edit', {'id': OrderItem.order_id}) }}">{{ OrderItem.order_id }}</a>
+        <a href="{{ url('admin_order_edit', {'id': OrderItem.order_id}) }}">{{ OrderItem.Order.order_code }}</a>
     </td>
     <td class="align-middle"><span>{{ OrderItem.price|price }}</span></td>
     <td class="align-middle"><span>{{ OrderItem.quantity }}</span></td>
