+++ src/Eccube/Resource/template/default/Mail/order.twig
@@ -17,7 +17,7 @@ file that was distributed with this source code.
 ************************************************
 
 ご注文番号：{{ Order.id }}
-ご注文コード: {{ Order.code }}
+ご注文コード: {{ Order.oder_code }}
 お支払い合計：{{ Order.payment_total|price}}
 お支払い方法：{{ Order.payment_method }}
 メッセージ：{{ Order.message }}
