+++ src/Eccube/Resource/template/default/Mypage/history.twig
@@ -44,12 +44,8 @@ file that was distributed with this source code.
                         <dd>{{ Order.create_date|date("Y/m/d H:i:s") }}</dd>
                     </div>
                     <div class="ec-definitions">
-                        <dt>{{ 'mypage.history.label.order_id'|trans}}</dt>
-                        <dd>{{ Order.id }}</dd>
-                    </div>
-                    <div class="ec-definitions">
-                        <dt>ご注文コード：</dt>
-                        <dd>{{ Order.code }}</dd>
+                        <dt>{{ 'mypage.history.label.order_code'|trans}}</dt>
+                        <dd>{{ Order.order_code }}</dd>
                     </div>
                     {% if BaseInfo.option_mypage_order_status_display %}
                         <div class="ec-definitions">
